# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Quote management REST API for the t-0 Network provider.",
        "title": "my-provider API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.agtpay.xyz",
    "basePath": "/",
    "paths": {
        "/api/v1/quotes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get stored quote snapshots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.SwaggerQuotesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/network": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get a real-time network quote",
                "parameters": [
                    {
                        "description": "Network quote request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quoteapi.SwaggerGetNetworkQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.NetworkQuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/pay-in": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Update pay-in quotes",
                "parameters": [
                    {
                        "description": "Pay-in quote groups",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quoteapi.SwaggerUpdateQuotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.PublishResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/pay-out": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Update pay-out quotes",
                "parameters": [
                    {
                        "description": "Pay-out quote groups",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quoteapi.SwaggerUpdateQuotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.PublishResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Push pay-out quotes to the network",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional idempotency key; if omitted, one is derived from the body hash",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Pay-out quote groups",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePayOutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePayOutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Publish quotes to the network",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.PublishResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/publish/pay-in": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Publish quotes to the network",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.PublishResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotes/publish/pay-out": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Publish quotes to the network",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.PublishResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/quoteapi.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "api.UpdatePayOutBand": {
            "type": "object",
            "properties": {
                "client_quote_id": {
                    "type": "string"
                },
                "max_amount_usd": {
                    "type": "string"
                },
                "rate": {
                    "type": "string"
                }
            }
        },
        "api.UpdatePayOutGroup": {
            "type": "object",
            "properties": {
                "bands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UpdatePayOutBand"
                    }
                },
                "currency": {
                    "type": "string"
                },
                "expiration_seconds": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                }
            }
        },
        "api.UpdatePayOutRequest": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UpdatePayOutGroup"
                    }
                }
            }
        },
        "api.UpdatePayOutResponse": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "description": "RFC3339",
                    "type": "string"
                },
                "bands_published": {
                    "type": "integer"
                },
                "expires_at": {
                    "description": "RFC3339",
                    "type": "string"
                },
                "groups_published": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "description": "always \"OK\"",
                    "type": "string"
                }
            }
        },
        "quoteapi.APIDecimal": {
            "type": "object",
            "properties": {
                "exponent": {
                    "type": "integer"
                },
                "unscaled": {
                    "type": "integer"
                }
            }
        },
        "quoteapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "quoteapi.NetworkProviderQuote": {
            "type": "object",
            "properties": {
                "executable": {
                    "type": "boolean"
                },
                "expiration": {
                    "type": "string"
                },
                "payOutAmount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "quoteId": {
                    "$ref": "#/definitions/quoteapi.NetworkQuoteID"
                },
                "rate": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "settlement": {
                    "$ref": "#/definitions/quoteapi.NetworkSettlement"
                }
            }
        },
        "quoteapi.NetworkQuoteFailure": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "quoteapi.NetworkQuoteID": {
            "type": "object",
            "properties": {
                "providerId": {
                    "type": "integer"
                },
                "quoteId": {
                    "type": "integer"
                }
            }
        },
        "quoteapi.NetworkQuoteResponse": {
            "type": "object",
            "properties": {
                "allQuotes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quoteapi.NetworkProviderQuote"
                    }
                },
                "result": {
                    "$ref": "#/definitions/quoteapi.NetworkQuoteResult"
                }
            }
        },
        "quoteapi.NetworkQuoteResult": {
            "type": "object",
            "properties": {
                "failure": {
                    "$ref": "#/definitions/quoteapi.NetworkQuoteFailure"
                },
                "success": {
                    "$ref": "#/definitions/quoteapi.NetworkQuoteSuccess"
                }
            }
        },
        "quoteapi.NetworkQuoteSuccess": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string"
                },
                "payOutAmount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "quoteId": {
                    "$ref": "#/definitions/quoteapi.NetworkQuoteID"
                },
                "rate": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "settlementAmount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                }
            }
        },
        "quoteapi.NetworkSettlement": {
            "type": "object",
            "properties": {
                "amount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "creditLimit": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "prefundingAmount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "totalUsed": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                }
            }
        },
        "quoteapi.PublishResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "published": {
                    "type": "boolean"
                }
            }
        },
        "quoteapi.SwaggerBand": {
            "type": "object",
            "properties": {
                "clientQuoteId": {
                    "type": "string"
                },
                "fix": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "maxAmount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "rate": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                }
            }
        },
        "quoteapi.SwaggerGetNetworkQuoteRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "$ref": "#/definitions/quoteapi.APIDecimal"
                },
                "amountType": {
                    "type": "string"
                },
                "payOutCurrency": {
                    "type": "string"
                },
                "payOutMethod": {
                    "type": "string"
                }
            }
        },
        "quoteapi.SwaggerQuoteGroup": {
            "type": "object",
            "properties": {
                "bands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quoteapi.SwaggerBand"
                    }
                },
                "currency": {
                    "type": "string"
                },
                "expiration": {
                    "type": "string"
                },
                "paymentMethod": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "quoteapi.SwaggerQuotesResponse": {
            "type": "object",
            "properties": {
                "payIn": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quoteapi.SwaggerQuoteGroup"
                    }
                },
                "payOut": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quoteapi.SwaggerQuoteGroup"
                    }
                }
            }
        },
        "quoteapi.SwaggerUpdateQuotesRequest": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quoteapi.SwaggerQuoteGroup"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}